Moved some misplaced code out of InputHandler, created a GRMouse class to handle manipulations of the mouse vector instead. My thinking is that InputHandler should be completely independent of GrappleRock, meaning we could reuse the class elsewhere without having to touch the code.

+++ src/util/Camera.java
@@ -79,11 +79,11 @@ public class Camera {
 
 	}
 	
-	public int getWorldXCoordinate(int screenX) {
+	public float getWorldXCoordinate(float screenX) {
 		return x + screenX;
 	}	
 	
-	public int getWorldYCoordinate(int screenY) {
+	public float getWorldYCoordinate(float screenY) {
 		return y + screenY;
 	}
 }
